requirements.txt
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-decouple
Pillow
drf-yasg
django-filter
django-jazmin